+++ lib/active_admin/inputs/filter_select_input.rb
@@ -11,8 +11,15 @@ module ActiveAdmin
       def input_name
         return method if seems_searchable?
 
-        name = method.to_s
-        name.concat '_id' if reflection
+        # Deal with has_many :through relationships in filters
+        # If the relationship is a HMT, we set the search logic to be something
+        # like :#{through_association}_#{end_association_id}.
+        if through_association?
+          name = [reflection.through_reflection.name, reflection.source_reflection.foreign_key].join('_')
+        else
+          name = method.to_s
+          name.concat '_id' if reflection
+        end
         name.concat multiple? ? '_in' : '_eq'
       end